LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/pseudo-element-anchor-expected.txt


FAIL ::before as anchor assert_equals: expected 100 but got 0
FAIL ::after as anchor assert_equals: expected 100 but got 0
